IBM_Bluemix/BlueSecure/src/requirements.txt
Flask==0.10.1
pycrypto
simple-crypt
python-swiftclient
python-keystoneclient